requirements.txt
altair==5.5.0
asttokens==3.0.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
bqplot==0.12.44
branca==0.8.1
cachetools==5.5.2
captcha==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colour==0.1.5
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
decorator==5.2.1
earthengine-api==1.5.4
eerepr==0.1.1
executing==2.2.0
extra-streamlit-components==0.1.71
folium==0.19.4
fonttools==4.56.0
future==1.0.0
geemap==0.35.1
geocoder==1.38.1
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.2
google-cloud-secret-manager==2.23.1
google-cloud-storage==3.1.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.68.0
grpc-google-iam-v1==0.14.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
httplib2==0.22.0
idna==3.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.19.2
ipython==8.31.0
ipython_pygments_lexers==1.1.1
ipytree==0.2.2
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-leaflet==0.19.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.28.0
numpy==2.2.3
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
plotly==6.0.0
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyogrio==0.10.0
pyparsing==3.2.1
pyperclip==1.9.0
pyproj==3.7.1
pyshp==2.3.1
python-box==7.3.2
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
ratelim==0.1.6
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
scooby==0.10.0
setuptools==75.8.0
shapely==2.0.7
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
streamlit==1.41.1
streamlit-authenticator==0.4.1
streamlit_folium==0.24.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
traitlets==5.14.3
traittypes==0.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
wheel==0.45.1
widgetsnbextension==4.0.13
xyzservices==2025.1.0
